+++ CHANGELOG.md
@@ -7,6 +7,11 @@ All notable changes to this project will be documented in this file.
 - New macro "Change Token Vision". It allows to apply vision types like Darksight, Sightless, and all the others to the selected tokens, using the new Foundry v10 token vision settings.
 - New Active Effect property, which allows character to reroll 1 on boon rolls
 - 2 new Active Effect properties: "Attack Boons/Banes - Spell" and "Attack Boons/Banes - Spell"
+- It is now possible to change the initiative order directly inside the combat tracker simply by dragging the combatants:
+  - Dragging with the Alt key pressed puts the dragged combatant below instead of above.
+  - GMs can drag all combatants, Players only owned combatants
+  - Dragging is limited to the initiative "speed slots". For example, characters with fast turns can freely move their order within other character with fast turns, but not creatures with fast or slow turns.
+  - Huge thanks to Mana#4176 for the code.
 ### Changed
 - Now measured templates dynamically target tokens even before placing them.
 - Auto targeting by measured templates is now enabled by default.